Update confstate in message before sending snapshot
Summary:
In the new version of raft library, a follower
doesn't apply a snapshot if it can't find itself in the
confstate of the snapshot. This can be problematic in cases
where a new node is added and the leader sends it a snapshot
taken before the confchange. This was caught by the test
`TestKronosSanityDeadNode` and this diff pulls in the fix
from the upstream.

More details in
[[ https://rubrik.atlassian.net/browse/CDM-430649?focusedCommentId=4311439 | CDM-430649 ]].

Test Plan:
- make test
- make acceptance

Reviewers: jj

Reviewed By: jj

Subscribers: jenkins.bot

JIRA Issues: CDM-430649, PHAB-5444

Differential Revision: https://phabricator.rubrik.com/D320492