Do not use seeds beyond Bootstrap.
Summary:
Currently, when a node applies a
snapshot, it fetches the current cluster metadata
from seeds to get the addresses for the nodes
present in the confstate. This diff replaces it
by fetching the addresses from gossip and then
querying the raft leader/other nodes to get
info about nodes which were added and then removed
before the snapshot.

Test Plan:
- make test
- make acceptance

Reviewers: grammar-police!, Sir.Alfred

JIRA Issues: CDM-450260

Differential Revision: https://phabricator.rubrik.com/D337577