Apply Valid snapshot on restart
Summary:
If a node crashes between saving the snapshot and updating the hard
state, the commit index in the hardstate will be less than the snapshot
index. However, this will cause problems on restart since we always
apply the newest snapshot  and initialize the first index of the
in-memory log to snapshotIndex - 1,  this breaks the invariant that
hard state's commit index must be between raftlog's commit
and applied indices

See
https://github.com/etcd-io/raft/blob/main/log.go#L92-L94

This diff fixes it by applying only valid snapshots (the snapshots which
have index <= Hardstate.commitIndex.

{P14558}

Test Plan:
- make test
- make acceptance

Reviewers: jj

Reviewed By: jj

Subscribers: Sir.Alfred, jj

JIRA Issues: CDM-442052, PHAB-7120

Differential Revision: https://phabricator.rubrik.com/D340062